movec tests from misbehave.c to exceptions.c and disabled them
