+++ tests/utest.c
@@ -98,5 +98,5 @@ UTEST_ADD_SIMPLE(access_basic);
 UTEST_ADD_SIMPLE(stat);
 UTEST_ADD_SIMPLE(fstat);
 
-UTEST_ADD_SIMPLE(userspace_cop_unusable);
-UTEST_ADD_SIMPLE(reserved_instruction);
+// UTEST_ADD_SIMPLE(userspace_cop_unusable);
+// UTEST_ADD_SIMPLE(reserved_instruction);
